reorganised directories
