+++ src/root/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 import {BrowserRouter as Router, Route} from "react-router-dom";
-import "./Styles/global.css";
+import "../Styles/global.css";
 
 const App = () => (
     <div className="App">
